Moved ShareActions and ShareStore into common folder. No functional changes.
